harden cycle 1 phase 2 pass 1: shared privilege helper, trace rule, ordering pin, doc corrections
